Uploaded "Incident Report and Analysis" PDF
Added the PDF guide "Incident Report and Analysis" to the repository. This guide covers essential concepts related to using the NIST cybersecurity framework and provides step-by-step instructions for responding to a security incident. The guide is part of my Google cybersecurity project portfolio and is intended to help others navigate and understand incident reports and analysis.